AIレビュー指摘を反映
