+++ src/app/login/page.tsx
@@ -1,11 +1,196 @@
-import { Suspense } from "react";
+"use client";
 
-import LoginForm from "./LoginForm";
+import { useState } from "react";
+import { useRouter } from "next/navigation";
+import Link from "next/link";
+
+import EyeIcon from "@/components/EyeIcon";
+import WindowPanel from "@/components/WindowPanel";
+import { useCurrentUser } from "@/contexts/CurrentUserContext";
+import { getApiBaseUrl } from "@/lib/apiConfig";
+import {
+  validateEmail,
+  validatePassword,
+  type ValidationErrors,
+} from "@/lib/validation";
+
+function validate(email: string, password: string): ValidationErrors {
+  const errors: ValidationErrors = {};
+  const emailError = validateEmail(email);
+  const passwordError = validatePassword(password);
+
+  if (emailError) errors.email = emailError;
+  if (passwordError) errors.password = passwordError;
+
+  return errors;
+}
 
 export default function LoginPage() {
+  const router = useRouter();
+  const { updateCurrentUser } = useCurrentUser();
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [showPassword, setShowPassword] = useState(false);
+  const [errors, setErrors] = useState<ValidationErrors>({});
+  const [isSubmitting, setIsSubmitting] = useState(false);
+
+  async function handleSubmit(e: React.FormEvent<HTMLFormElement>) {
+    e.preventDefault();
+
+    const validationErrors = validate(email, password);
+    if (Object.keys(validationErrors).length > 0) {
+      setErrors(validationErrors);
+      return;
+    }
+
+    setIsSubmitting(true);
+    setErrors({});
+
+    try {
+      const res = await fetch(`${getApiBaseUrl()}/api/auth/login`, {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
+        credentials: "include",
+        body: JSON.stringify({ email, password }),
+      });
+
+      if (!res.ok) {
+        const data: { message?: string } = await res.json().catch(() => ({}));
+        setErrors({
+          general:
+            data.message ?? "メールアドレスまたはパスワードが正しくありません",
+        });
+        return;
+      }
+
+      const data: { id: number; name: string } = await res.json();
+
+      localStorage.setItem("userId", String(data.id));
+      localStorage.setItem("userName", data.name);
+
+      updateCurrentUser({
+        id: data.id,
+        name: data.name,
+      });
+
+      router.push("/home");
+    } catch {
+      setErrors({ general: "通信エラーが発生しました" });
+    } finally {
+      setIsSubmitting(false);
+    }
+  }
+
   return (
-    <Suspense fallback={null}>
-      <LoginForm />
-    </Suspense>
+    <main className="flex flex-1 items-center justify-center overflow-auto">
+      <div className="[&>section]:min-h-[500px] [&>section]:min-w-[535px]">
+        <WindowPanel>
+          <h1 className="transform-[perspective(320px)_rotateX(12deg)_skewX(-8deg)_scaleY(1.08)] text-4xl font-bold tracking-wide text-white drop-shadow-[4px_4px_0_#64748b]">
+            Type<span className="text-yellow-300">★</span>Clash
+          </h1>
+
+          <p className="mt-2 whitespace-nowrap text-base tracking-widest text-slate-300">
+            リアルタイム対戦型タイピングアプリ
+          </p>
+
+          <form
+            onSubmit={handleSubmit}
+            noValidate
+            className="mt-8 flex w-full max-w-sm flex-col gap-4"
+          >
+            {/* 通信エラー */}
+            <p
+              role={errors.general ? "alert" : undefined}
+              className={`min-h-[20px] text-left text-base font-semibold ${errors.general ? "text-red-400" : "text-transparent"}`}
+            >
+              {errors.general ?? "\u00A0"}
+            </p>
+
+            {/* メールアドレス */}
+            <div className="flex flex-col gap-1">
+              <label
+                htmlFor="email"
+                role={errors.email ? "alert" : undefined}
+                className={`min-h-[20px] text-left text-sm ${errors.email ? "text-red-300" : "text-slate-200"}`}
+              >
+                {errors.email ?? "メールアドレス"}
+              </label>
+
+              <input
+                id="email"
+                type="email"
+                value={email}
+                onChange={(e) => setEmail(e.target.value)}
+                autoComplete="email"
+                placeholder="*****@outlook.jp"
+                className="border border-slate-400 bg-[#0d1b3e]/80 px-4 py-2 text-lg text-white outline-none placeholder:text-slate-400 focus:border-slate-100"
+              />
+            </div>
+
+            {/* パスワード */}
+            <div className="flex flex-col gap-1">
+              <label
+                htmlFor="password"
+                role={errors.password ? "alert" : undefined}
+                className={`min-h-[20px] text-left text-sm ${errors.password ? "text-red-300" : "text-slate-200"}`}
+              >
+                {errors.password ?? "パスワード"}
+              </label>
+
+              <div className="relative">
+                <input
+                  id="password"
+                  type={showPassword ? "text" : "password"}
+                  value={password}
+                  onChange={(e) => setPassword(e.target.value)}
+                  autoComplete="current-password"
+                  className="w-full border border-slate-400 bg-[#0d1b3e]/80 px-4 py-2 pr-12 text-lg text-white outline-none placeholder:text-slate-400 focus:border-slate-100"
+                />
+                <button
+                  type="button"
+                  onClick={() => setShowPassword((v) => !v)}
+                  className="absolute right-3 top-1/2 -translate-y-1/2 text-slate-400 hover:text-slate-200"
+                  aria-label={
+                    showPassword ? "パスワードを隠す" : "パスワードを表示する"
+                  }
+                >
+                  <EyeIcon isHidden={!showPassword} />
+                </button>
+              </div>
+            </div>
+
+            <p className="-mt-1.5 text-left text-sm">
+              <Link href="/password-reset/request">
+                <span className="text-slate-300 underline-offset-2 transition-colors hover:text-sky-400 hover:underline">
+                  パスワードを忘れた方はこちら
+                </span>
+              </Link>
+            </p>
+
+            {/* ログインボタン */}
+            <button
+              type="submit"
+              disabled={isSubmitting}
+              className="mt-1 border border-slate-400 bg-[#0d1b3e]/60 py-2 text-lg font-bold tracking-[0.03em] text-white hover:border-yellow-200 hover:text-yellow-200 disabled:cursor-not-allowed disabled:opacity-50"
+            >
+              {isSubmitting ? "読み込み中..." : "ログイン"}
+            </button>
+
+            <p className="mt-2.5 text-center text-[1.15rem]">
+              <Link href="/register">
+                <span className="text-blue-300 underline-offset-2 transition-colors hover:text-sky-400 hover:underline">
+                  新規登録はこちら
+                </span>
+              </Link>
+            </p>
+          </form>
+          <Link href="/" className="mt-6 self-start">
+            <span className="text-lg transition-colors hover:text-yellow-200">
+              ◀ トップページへ
+            </span>
+          </Link>
+        </WindowPanel>
+      </div>
+    </main>
   );
 }
